room updates availability automatically

+++ C#/Application/Logic/PatientLogic.cs
@@ -1,4 +1,5 @@
-﻿using Application.DaoInterfaces;
+﻿using System.Security.AccessControl;
+using Application.DaoInterfaces;
 using Application.LogicInterfaces;
 using Domain.DTOs;
 using Domain.Models;
@@ -15,16 +16,35 @@ public class PatientLogic : IPatientLogic
         this.roomDao = roomDao;
         this.patientDao = patientDao;
     }
+
     public async Task<Patient> CreateAndAddToRoomAsync(int roomId, PatientCreationDto dto)
     {
         ValidatePatient(dto);
+        Room? room = await roomDao.GetById(roomId);
+        if (room == null)
+            throw new Exception($"Room with id {roomId} not found");
+        if (room.Capacity <= room.Patients.Count)
+        {
+            throw new Exception("Room is full. Cannot add more patients!");
+        }
+
         Patient toCreate = new Patient()
         {
             Name = dto.Name
         };
-        Patient created = await roomDao.CreateAndAddToRoomAsync(roomId, toCreate);
+        Patient created = await patientDao.CreateAsync(toCreate);
+        room.Patients.Add(created);
+  
+        int capacity = room.Capacity;
+        int actual = room.Patients.Count;
+        if (capacity == actual)
+        {
+            room.Availability = "Occupied";
+        } 
+        await roomDao.RoomUpdateAsync(room);
         return created;
     }
+
     private void ValidatePatient(PatientCreationDto dto)
     {
         if (dto.Name == "")
@@ -42,9 +62,30 @@ public class PatientLogic : IPatientLogic
         {
             throw new Exception($"Patient with ID {id} was not found!");
         }
-
-
+        Room? room = await roomDao.GetRoomWithPatientId(id);
+        if (room == null)
+            throw new Exception("Room not found");
+        bool wasFull = room.Availability == "Occupied";
         await patientDao.DeleteAsync(id);
+        
+        if (wasFull)
+        {
+            room.Availability = "Available";
+        } 
+        await roomDao.RoomUpdateAsync(room);
+    }
+    private bool IsRoomNotFull(int roomId)
+    {
+        bool ok = false;
+        Room? room = roomDao.GetByIdToUpdateAsync(roomId).Result;
+        if (room != null)
+        {
+            int capacity = room.Capacity;
+            int actual = room.Patients.Count;
+            if (actual == capacity - 1)
+                ok = true;
+        }
+        return ok;
     }
     
     public async Task<Patient?> GetByIdAsync(int id)
